move webapp dataSafe check to utils module
